deactivate shape from pan and zoom dragmodes  - move shape activation & deactivation functions to shape
