+++ src/plots/cartesian/handle_outline.js
@@ -9,46 +9,6 @@
 
 'use strict';
 
-var Registry = require('../../registry');
-
-
-function activateShape(gd, path, drawShapes) {
-    var element = path.node();
-    var id = +element.getAttribute('data-index');
-    gd._fullLayout._activeShapeIndex = id;
-    if(id >= 0) drawShapes(gd);
-}
-
-function deactivateShape(gd) {
-    clearOutlineControllers(gd);
-
-    delete gd._fullLayout._activeShapeIndex;
-    Registry.call('_guiRelayout', gd, {
-        shapes: (gd.layout || {}).shapes || []
-    });
-}
-
-function eraseActiveShape(gd) {
-    clearOutlineControllers(gd);
-
-    var id = gd._fullLayout._activeShapeIndex;
-    var shapes = (gd.layout || {}).shapes || [];
-    if(id < shapes.length) {
-        var allShapes = [];
-        for(var q = 0; q < shapes.length; q++) {
-            if(q !== id) {
-                allShapes.push(shapes[q]);
-            }
-        }
-
-        delete gd._fullLayout._activeShapeIndex;
-
-        Registry.call('_guiRelayout', gd, {
-            shapes: allShapes
-        });
-    }
-}
-
 function clearOutlineControllers(gd) {
     var zoomLayer = gd._fullLayout._zoomlayer;
     if(zoomLayer) {
@@ -69,9 +29,6 @@ function clearSelect(gd) {
 }
 
 module.exports = {
-    activateShape: activateShape,
-    deactivateShape: deactivateShape,
-    eraseActiveShape: eraseActiveShape,
     clearOutlineControllers: clearOutlineControllers,
     clearSelect: clearSelect
 };
